Add gamelogic.js file with GameLogic module
